+++ PiesGenericPatch/ModFolderContent/CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.3.0
+
+- Fix group sign-up posters not rendering text correctly
+- Fix signatures not being signatures on documents, while leaving sales ledgers alone
+
 # 1.2.0
 
 - Fix citizen needs (bladder especially) increasing many times faster than intended while they interact or talk to the player